made the friends side bare a little simpler and ccleaner
